feat: add health endpoint, custom 404, and OG meta tags
Production readiness improvements:
- /api/health endpoint with DB check, uptime, and stats
- Custom 404 page matching app design
- Open Graph + Twitter Card meta tags on listing pages
- Base OG metadata in root layout

Closes #40
